requirements.txt
Django==4.2.16
djangorestframework==3.15.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.7.0
django-environ==0.11.2
djangorestframework-simplejwt==5.3.1
